style fonts left, ts connect button and menu left
