+++ src/scss/styles.scss
@@ -1,0 +1,1 @@
+@import "../components/timerdigitalstyle/timerdigital";
